Fixed head address max value

+++ cs/src/core/Index/FasterLog/FasterLog.cs
@@ -135,7 +135,7 @@ namespace FASTER.core
             if (logSettings.ReadOnlyMode)
             {
                 readOnlyMode = true;
-                allocator.HeadAddress = int.MaxValue;
+                allocator.HeadAddress = long.MaxValue;
             }
 
             Restore(out RecoveredIterators);
